Added image delete functionality with password verification

+++ server/controllers/imageController.js
@@ -1,3 +1,5 @@
+//depedencies
+const bcrypt = require('bcrypt');
 //internal imports
 const Image = require('./../models/Image');
 const User = require('./../models/User');
@@ -15,14 +17,24 @@ const addImage = async function (req, res, next) {
 
 const deleteImage = async function (req, res, next) {
    try {
-      const deletedImage = await Image.findByIdAndRemove({ _id: req.body.id });
-      await User.updateOne(
-         { _id: req.user.id },
-         {
-            $pull: { images: deletedImage._id },
+      const user = await User.findOne({ _id: req.user.id });
+      if (user && user._id) {
+         const isValidPassword = await bcrypt.compare(req.body.password, user.password);
+         if (isValidPassword) {
+            const deletedImage = await Image.findByIdAndRemove({ _id: req.body.id });
+            await User.updateOne(
+               { _id: req.user.id },
+               {
+                  $pull: { images: deletedImage._id },
+               }
+            );
+            res.status(200).json({ status: 'success', message: 'Image deleted successfully' });
+         } else {
+            next(new Error('Your entered password that does not match with the logged in user password'));
          }
-      );
-      res.status(200).json({ status: 'success', message: 'Image deleted successfully' });
+      } else {
+         next(new Error('User not found'));
+      }
    } catch (error) {
       next(new Error('There was error while deleteing the image'));
    }
